refactor(website): extract workflow animation hook and remove DrawEdge id prop
Extract useWorkflowAnimation hook to separate timer side-effect from
WorkflowVisual component. Remove redundant id prop from DrawEdge since
React's key on the component handles reconciliation.
